+++ SpermListTest1/data/entites/sperm.cs
@@ -10,18 +10,18 @@ namespace SpermListTest1.data.entites
         public string NAAB_CODE { get; set; }
         public string NAME { get; set; }
         public double ICC { get; set; }
-        public int LNM { get; set; }
-        public int LFM { get; set; }
-        public int MILK { get; set; }
+        public double LNM { get; set; }
+        public double LFM { get; set; }
+        public double MILK { get; set; }
         public double FAT { get; set; }
-        public int PRO { get; set; }
+        public double PRO { get; set; }
         public double SCE { get; set; }
         public double PL { get; set; }
         public double DPR { get; set; }
         public double PTAT { get; set; }
         public double UDC { get; set; }
         public double FLC { get; set; }
-        public int TPI { get; set; }
+        public double TPI { get; set; }
         public string SIRE { get; set; }
         public string MGS { get; set; }
     }
